Moved stylesheet to seperate "styles.js" file
